requirements.txt
streamlit==1.41.0
mysql-connector-python==9.2.0
openai>=1.52.0
bcrypt==4.2.1
python-dotenv==1.0.1
Authlib==1.4.1
requests==2.32.3
plotly==6.0.0
httpx==0.28.1
